Move SyncAlgorithm, CollectionSyncAlgorithm and PropfindReportAlgorithm to sync.algorithm package
