+++ core/src/main/kotlin/at/bitfire/davdroid/sync/algorithm/CollectionSyncAlgorithm.kt
@@ -2,7 +2,7 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid.sync
+package at.bitfire.davdroid.sync.algorithm
 
 import at.bitfire.dav4jvm.Error
 import at.bitfire.dav4jvm.ktor.exception.HttpException
@@ -10,6 +10,7 @@ import at.bitfire.dav4jvm.property.webdav.SyncToken
 import at.bitfire.dav4jvm.property.webdav.WebDAV
 import at.bitfire.davdroid.resource.LocalCollection
 import at.bitfire.davdroid.resource.SyncState
+import at.bitfire.davdroid.sync.MultiResponseCallback
 import java.util.logging.Logger
 
 /**
@@ -20,7 +21,7 @@ class CollectionSyncAlgorithm(
     private val context: Context
 ) : SyncAlgorithm {
 
-    /** Operations this algorithm needs from the owning [SyncManager]. */
+    /** Operations this algorithm needs from the owning [at.bitfire.davdroid.sync.SyncManager]. */
     class Context(
         val localCollection: LocalCollection<*>,
         val resetPresentRemotely: () -> Unit,
