+++ core/src/main/kotlin/at/bitfire/davdroid/sync/algorithm/SyncAlgorithm.kt
@@ -2,15 +2,16 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid.sync
+package at.bitfire.davdroid.sync.algorithm
 
 import at.bitfire.dav4jvm.Property
 import at.bitfire.dav4jvm.ktor.MultiStatusItem
 import at.bitfire.davdroid.resource.SyncState
+import at.bitfire.davdroid.sync.MultiResponseCallback
 import kotlinx.coroutines.flow.Flow
 
 /**
- * A pluggable synchronization algorithm, run by [SyncManager] once it has been determined
+ * A pluggable synchronization algorithm, run by [at.bitfire.davdroid.sync.SyncManager] once it has been determined
  * that a sync run is necessary.
  */
 interface SyncAlgorithm {
